--- Apackage.json
@@ -1,19 +1,0 @@
-{
-  "name": "googlecalendarapi",
-  "version": "1.0.0",
-  "description": "",
-  "main": "app.js",
-  "scripts": {
-    "start": "node src/app.js",
-    "dev": "env-cmd -f ./config/dev.env node src/app.js"
-  },
-  "author": "Nadia Berry",
-  "license": "MIT",
-  "dependencies": {
-    "@sendgrid/mail": "^7.4.0",
-    "ejs": "^3.1.5",
-    "env-cmd": "^10.1.0",
-    "express": "^4.17.1",
-    "googleapis": "^39.2.0"
-  }
-}
