+++ package.json
@@ -4,6 +4,8 @@
   "description": "Hack-a-thon Project by Aaron Moua, Fatima Cruz, Nadia Berry &amp; Jared Johnson",
   "main": "index.js",
   "scripts": {
+    "start": "node src/app.js",
+    "dev": "env-cmd -f ./config/dev.env node src/app.js",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "repository": {
@@ -17,10 +19,12 @@
   },
   "homepage": "https://github.com/jjohn513/How-Was-School#readme",
   "dependencies": {
+    "@sendgrid/mail": "^7.4.0",
     "body-parser": "^1.19.0",
     "dotenv": "^8.2.0",
     "ejs": "^3.1.5",
     "express": "^4.17.1",
+    "googleapis": "^66.0.0",
     "morgan": "^1.10.0",
     "request": "^2.88.2"
   }
