+++ views/calendar.ejs
@@ -99,41 +99,26 @@ body {font-family: Verdana, sans-serif;}
 
 
     <!-- Calendar Form -->
-    <div class="calendarForm">
-      <form action="/action_page.php">
-        <label for="eventName">Event Name</label>
-        <input type="text" id="eventName" >
-        
-
-        <form action="/action_page.php">
-      <label for="date">Event Date</label>
-      <input type="date" id="date" name="date">
-    </form>
-    
-    
-    <form action="/action_page.php">
-      <label for="startTime">Start Time:</label>
-      <input type="time" id="time" name="time">
-    </form>
-    
-
-    <form action="/action_page.php">
-      <label for="time">End Time:</label>
-      <input type="time" id="time" name="time">
-    </form>
-  
-        
-    <form action="/action_page.php">
-    <label for="description">Event Description</label>
-    <textarea id="description" name="description" rows="4" cols="50">
-      </textarea>
-      
-        <input type="submit" value="Submit">
+    <div class="form">
+      <form action="http://localhost:3025/" method="POST">
+        <a class="button" target="_blank" rel="noopener noreferrer" href="https://accounts.google.com/o/oauth2/v2/auth/oauthchooseaccount?access_type=offline&scope=https%3A%2F%2Fwww.googleapis.com%2Fauth%2Fcalendar.readonly&response_type=code&client_id=808233503594-c3isbglgijq2l1op2hviv9ojqg0sms5m.apps.googleusercontent.com&redirect_uri=urn%3Aietf%3Awg%3Aoauth%3A2.0%3Aoob&flowName=GeneralOAuthFlow">Sign In</a>
+         <span>You need to Sign In first</span><br><br>
+        <label for="token">Paste the copied token here</label>
+        <input name="token" type="text" placeholder="must be like this Jw65AJg05xiOGmkCld...">
+        <input type="submit" value="Get my events"></input>
+      </form>
+      <hr>
+      <form action="http://localhost:3025/events" method="POST">
+        <label for="attendee email">Attendee email</label>
+        <input type="email" name="to" placeholder="send email to...">
+        <label for="event title">Event Title</label>
+        <input type="text" name="summary" placeholder="the title of event...">
+        <label for="body">Event body</label>
+        <input type="text" name="description" placeholder="the message body...">
+        <input type="submit" value="Send Event Invite">
       </form>
     </div>
     
-<!-- footer partial.  -->
-<%- include('partials/footer.ejs') %>
 
 </body>
