+++ views/events.ejs
@@ -3,7 +3,7 @@
 <head>
   <meta charset="UTF-8">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <link rel="stylesheet" href="src/public/css/styles.css">
+  <link rel="stylesheet" href="/public/css/calendar.css">
   <title>Google Events</title>
 </head>
 <body>
